Adding Test file to show case my new learned ability to make commits with structured messages
I learned that the structure that is best is the following

Subject = concise summary of what happened

Body = more detailed explanation
     - What is now different than before?
     - What's the reason for the change?
     - Is there anything to out for / anything particularly remarkable

Furthermore, these changes are constituted to add this file as a note for my learning proccess. There are no changes besides  adding of this file. Nothing to look out for.